% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{answer_calculate_sum}
\alias{answer_calculate_sum}
\title{Tính các giá trị tổng và tích giữa X và Y.}
\usage{
answer_calculate_sum(data, score = 1)
}
\description{
Hàm này tính tất cả các giá trị tổng giữa x và y
}
