% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke_dapan.R
\name{answer_correlation}
\alias{answer_correlation}
\title{Hàm này in ra đáp án cho bài toán tính hệ số tương quan}
\usage{
answer_correlation(x, y)
}
\description{
Hàm này in ra đáp án cho bài toán tính hệ số tương quan
}
