% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{answer_custom}
\alias{answer_custom}
\title{Tạo đáp án thủ công}
\usage{
answer_custom(templates, scores, data)
}
\arguments{
\item{templates:}{Một vector chứa mẫu đáp án}

\item{scores:}{Một vector chứa điểm số cho từng mục.}

\item{data:}{Một list chứa dữ liệu để thêm vào template.}
}
\description{
Hàm này dùng để tạo đáp án tùy ý, theo nhu cầu của người dùng.
}
