% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{answer_estimate_var}
\alias{answer_estimate_var}
\title{Trình bày bài toán ước lượng khoảng cho phương sai.}
\usage{
answer_estimate_var(data, answer, score = c(0.5, 1.5), round_digits = 4)
}
\description{
Hàm này in ra đáp án cho bài toán ước lượng phương sai
conclusion là lời giải của bài toán.
}
