% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{answer_linear_regression}
\alias{answer_linear_regression}
\title{Hàm này in ra đáp án cho bài toán hồi quy tuyến tính đơn và tìm giá trị dự báo của Y}
\usage{
answer_linear_regression(data, score = c(1, 0.5))
}
\description{
Hàm này in ra đáp án cho bài toán hồi quy tuyến tính đơn và tìm giá trị dự báo của Y
}
