% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{answer_linear_regression_predict}
\alias{answer_linear_regression_predict}
\title{Hàm này in ra đáp án cho bài toán hồi quy tuyến tính đơn và tìm giá trị dự báo của Y}
\usage{
answer_linear_regression_predict(data, conclusion, value_unit, score)
}
\description{
Hàm này in ra đáp án cho bài toán hồi quy tuyến tính đơn và tìm giá trị dự báo của Y
}
