% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{answer_mean}
\alias{answer_mean}
\title{Tính giá trị trung bình}
\usage{
answer_mean(data, answer, score = 1, round_digits = 4)
}
\arguments{
\item{data:}{là 1 vector dữ liệu}

\item{answer:}{câu trả lời của bài toán}

\item{score:}{Đáp số cho câu hỏi này.}
}
\description{
Hàm in đáp án bài toán tính giá trị trung bình
}
