% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{answer_merge}
\alias{answer_merge}
\title{Ghép đáp án}
\usage{
answer_merge(templates)
}
\arguments{
\item{templates}{Một vector chứa các template cần ghép.}
}
\description{
Hàm này dùng để ghép đáp án từ nhiều hàm có sẵn trở thành 1 đáp án duy nhất
}
