% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{answer_sd}
\alias{answer_sd}
\title{Tính giá trị độ lệch chuẩn}
\usage{
answer_sd(data, answer, with_mean = TRUE, score = 1, round_digits = 4)
}
\description{
Hàm này in đáp án cho bài toán tính giá trị độ lệch chuẩn
}
