% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke_dapan.R
\name{answer_test_2_mean_t}
\alias{answer_test_2_mean_t}
\title{Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)}
\usage{
answer_test_2_mean_t(n1, n2, mean1, mean2, s1, s2, alpha, mode = "neq")
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)
}
