% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{answer_test_2_prop}
\alias{answer_test_2_prop}
\title{Trình bày bài toán KĐGT về so sánh 2 tỷ lệ.}
\usage{
answer_test_2_prop(data, score = c(0.5, 1, 0.5), round_digits = 4)
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về so sánh 2 tỷ lệ.
}
