% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{answer_test_independent}
\alias{answer_test_independent}
\title{Trình bày bài toán kiểm định tính độc lập.}
\usage{
answer_test_independent(
  data,
  h0,
  row_names,
  col_names,
  conclusion_h0 = "",
  conclusion_h1 = "",
  score = c(0.5, 1, 0.5),
  round_digits = 4
)
}
\description{
Hàm này in ra đáp án cho bài toán kiểm định tính độc lập.
}
