% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_data}
\alias{generate_data}
\title{Hàm này dùng để gọi các hàm trong thư viện thongke, trả về các dữ liệu cần thiết
cho việc tạo đáp án}
\usage{
generate_data(func, params)
}
\description{
Hàm này dùng để gọi các hàm trong thư viện thongke, trả về các dữ liệu cần thiết
cho việc tạo đáp án
}
