% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{get_conclusion}
\alias{get_conclusion}
\title{Hàm này đưa ra kết luận cho bài toán kiểm định giả thiết}
\usage{
get_conclusion(rejected, operator)
}
\description{
Hàm này đưa ra kết luận cho bài toán kiểm định giả thiết
}
