% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{init_template}
\alias{init_template}
\title{Hàm này dùng để tạo template cho việc tạo đề thi bằng thongke.dapan}
\usage{
init_template()
}
\description{
Hàm này dùng để tạo template cho việc tạo đề thi bằng thongke.dapan
}
