% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke_dapan.R
\name{render_template}
\alias{render_template}
\title{Hàm này dùng thư viện whisker để render file}
\usage{
render_template(file_name, var_list)
}
\description{
Hàm này dùng thư viện whisker để render file
}
