% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke_dapan.R
\name{answer_estimate_mean_t}
\alias{answer_estimate_mean_t}
\title{Hàm này in ra đáp án cho bài toán ước lượng giá trị trung bình (dùng phân bố Student)}
\usage{
answer_estimate_mean_t(n, mean, s, alpha)
}
\description{
Hàm này in ra đáp án cho bài toán ước lượng giá trị trung bình (dùng phân bố Student)
}
