% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{answer_estimate_prop}
\alias{answer_estimate_prop}
\title{Trình bày bài toán ước lượng cho tỷ lệ.}
\usage{
answer_estimate_prop(data, conclusion, score = c(0.5, 1.5), round_digits = 4)
}
\description{
Hàm này in ra đáp án cho bài toán ước lượng cho tỷ lệ
}
