% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{answer_estimate_var}
\alias{answer_estimate_var}
\title{Hàm này in ra đáp án cho bài toán ước lượng phương sai}
\usage{
answer_estimate_var(n, s, alpha)
}
\description{
Hàm này in ra đáp án cho bài toán ước lượng phương sai
}
