% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{answer_linear_regression_predict}
\alias{answer_linear_regression_predict}
\title{Trình bày bài toán dự đoán dựa trên mô hình hồi quy.}
\usage{
answer_linear_regression_predict(
  data,
  answer,
  value_unit,
  score = 0.5,
  round_digits = 4
)
}
\description{
Hàm này in ra đáp án cho bài toán tìm giá trị dự báo của Y dựa trên mô hình hồi quy tuyến tính.
}
