% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{answer_sample_size_mean}
\alias{answer_sample_size_mean}
\title{Trình bày bài toán xác định kích thước mẫu (TH cho giá trị trung bình)}
\usage{
answer_sample_size_mean(
  data,
  sd_symbol,
  conclusion = function(value) {     return(sprintf("\%d", value)) },
  score = c(0.5, 1, 0.5),
  round_digits = 4
)
}
\description{
Hàm này in ra đáp án cho bài toán xác định kích thước mẫu (với giá trị trung bình)
conclusion là một hàm in ra kết quả cuối cùng của bài toán, với 1 tham số chính là kết quả của phép tính.
}
