% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{answer_sample_size_mean}
\alias{answer_sample_size_mean}
\title{Hàm này in ra đáp án cho bài toán xác định kích thước mẫu (với giá trị trung bình)}
\usage{
answer_sample_size_mean(data, sd_symbol)
}
\description{
Hàm này in ra đáp án cho bài toán xác định kích thước mẫu (với giá trị trung bình)
}
