% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{answer_test_2_mean_norm}
\alias{answer_test_2_mean_norm}
\title{Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)}
\usage{
answer_test_2_mean_norm(
  n1,
  n2,
  mean1,
  mean2,
  sigma1,
  sigma2,
  alpha,
  mode = "neq"
)
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)
}
