% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{answer_test_2_mean_t}
\alias{answer_test_2_mean_t}
\title{Trình bày bài toán KĐGT về so sánh 2 giá trị trung bình (phân bố Student)}
\usage{
answer_test_2_mean_t(
  data,
  conclusion_h0,
  conclusion_h1,
  score = c(0.5, 1, 0.5),
  round_digits = 4
)
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về so sánh 2 giá trị trung bình (phân bố Student)
}
