% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{answer_test_goodness_of_fit}
\alias{answer_test_goodness_of_fit}
\title{Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)}
\usage{
answer_test_goodness_of_fit(data, h0, col_names, score = c(0.5, 1, 0.5))
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)
}
