% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{answer_test_prop}
\alias{answer_test_prop}
\title{Trình bày bài toán KĐGT về tỷ lệ}
\usage{
answer_test_prop(
  data,
  conclusion_h0,
  conclusion_h1,
  score = c(0.5, 1, 0.5),
  round_digits = 4
)
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về tỷ lệ
}
