% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke_dapan.R
\name{answer_test_prop}
\alias{answer_test_prop}
\title{Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)}
\usage{
answer_test_prop(n, f, p_0, alpha, mode = "neq")
}
\description{
Hàm này in ra đáp án cho bài toán KĐGT về giá trị trung bình (phân bố chuẩn)
}
