% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{answer_var}
\alias{answer_var}
\title{Tính giá trị phương sai}
\usage{
answer_var(data, answer, with_mean = TRUE, score = 1, round_digits = 4)
}
\description{
Hàm này in đáp án cho bài toán tính phương sai
}
