% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke_dapan.R
\name{get_conclusion}
\alias{get_conclusion}
\title{Hàm này đưa ra kết luận cho bài toán kiểm định giả thiết}
\usage{
get_conclusion(test, c)
}
\description{
Hàm này đưa ra kết luận cho bài toán kiểm định giả thiết
}
